Update System.r
Add the ability to specify the number of exposed and infected states to a seir model.  If the "seir" option is chosen for the sckm function, an additional argument n can be specified.  n should be a named vector with elements ne, the number of exposed states, and ni, the number of infected states.  Should have at least 1 infected state.

Example:  system1 <- sckm("seir", n=c(ne=3,ni=3))